protect video detail routes

+++ controllers/videoController.js
@@ -109,12 +109,17 @@ export const deleteVideo = async (req, res) => {
     { _id: false, fileUrl: true }
   );
   try {
-    await Video.findOneAndDelete({ _id: id });
-    await fs.unlink(filePath.fileUrl, (err) => {
-      if (err) {
-        console.log(err);
-      }
-    });
+    const video = await Video.findById(id);
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndDelete({ _id: id });
+      await fs.unlink(filePath.fileUrl, (err) => {
+        if (err) {
+          console.log(err);
+        }
+      });
+    }
   } catch (error) {
     console.log(error);
   }
